system-tray The class that launches the application tray via the system tray has been moved into its own file.
