+++ src/tkinter_version/renderer.py
@@ -1,28 +1,5 @@
 import tkinter as tk
 from colour_converter import ColourConverter
-from PySide2 import QtWidgets, QtGui
-import os
-import sys
-
-class SystemTrayIcon(QtWidgets.QSystemTrayIcon):
-    def __init__(self, icon, parent=None):
-        QtWidgets.QSystemTrayIcon.__init__(self, icon, parent)
-        self.setToolTip("Chrispy Colour Palette")
-        menu = QtWidgets.QMenu(parent)
-
-        # Context Menu
-        open_colour_palette = menu.addAction("Open palette")
-        open_colour_palette.triggered.connect(self.open_colour_palette_app)
-        
-        _exit = menu.addAction("Exit")
-        _exit.triggered.connect(lambda: sys.exit())
-
-        menu.addSeparator()
-        self.setContextMenu(menu)
-        self.activated.connect(self.open_colour_palette_app)
-
-    def open_colour_palette_app(self):
-        Renderer().render()
 
 class Renderer():
     """
@@ -70,10 +47,11 @@ class Renderer():
         """
         self.root.title("Chrispy Colour Palette")
         self.root.iconbitmap("favicon.ico")
+        self.root.resizable(False, False)
         self.root.configure(
             bg="#1D1D1D", 
             padx="15", 
-            pady="25"
+            pady="25",
         )
 
     def render_hex_frame(self):
@@ -239,13 +217,3 @@ class Renderer():
             )
             self.hex_input.delete(0, "end")
             self.hex_input.insert(0, hex_value)
-            
-def main():
-    app = QtWidgets.QApplication(sys.argv)
-    widget = QtWidgets.QWidget()
-    tray_icon = SystemTrayIcon(QtGui.QIcon("favicon.ico"), widget)
-    tray_icon.show()
-    sys.exit(app.exec_())
-
-if __name__ == "__main__":
-    main()
